# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "初始化路由",
        "title": "API",
        "contact": {
            "name": "wty92911",
            "url": "https://github.com/wty92911",
            "email": "2066946886@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/food": {
            "post": {
                "description": "创建食品，返回创建好的食品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food"
                ],
                "summary": "创建食品",
                "parameters": [
                    {
                        "description": "创建食品请求参数",
                        "name": "food",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Food"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Food"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/foods": {
            "get": {
                "description": "根据分类ID获取食物列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food"
                ],
                "summary": "根据分类获取食物",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Food"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "github_com_wty92911_GoPigKit_internal_model.Food": {
            "type": "object",
            "required": [
                "category_id",
                "created_by",
                "image_urls",
                "price",
                "title"
            ],
            "properties": {
                "category_id": {
                    "description": "食品分类ID，外键",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdUser": {
                    "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                },
                "created_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "desc": {
                    "description": "食品描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_urls": {
                    "description": "食品图片",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "description": "食品价格",
                    "type": "integer"
                },
                "title": {
                    "description": "食品名称",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wty92911_GoPigKit_internal_model.User": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "family_id": {
                    "description": "user创建时可以不指定family_id，等后面再加入一个家庭",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "open_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}